Complete all tasks for "Biztonság" homework.

+++ src/main/java/hu/cubix/hr/akos0012/service/timeOffRequest/TimeOffRequestService.java
@@ -1,6 +1,5 @@
 package hu.cubix.hr.akos0012.service.timeOffRequest;
 
-import hu.cubix.hr.akos0012.dto.TimeOffRequestFilterSortDTO;
 import hu.cubix.hr.akos0012.model.Employee;
 import hu.cubix.hr.akos0012.model.RequestStatus;
 import hu.cubix.hr.akos0012.model.TimeOffRequest;
@@ -12,6 +11,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
@@ -51,9 +52,14 @@ public class TimeOffRequestService {
         return startDate.isBefore(endDate);
     }
 
-    public TimeOffRequest create(long employeeID, LocalDateTime startDate, LocalDateTime endDate) {
-        Optional<Employee> employee = employeeRepository.findById(employeeID);
+    private String getCurrentUsername() {
+        return SecurityContextHolder.getContext().getAuthentication().getName();
+    }
+
+    public TimeOffRequest create(LocalDateTime startDate, LocalDateTime endDate) {
+        Optional<Employee> employee = employeeRepository.findByUsername(getCurrentUsername());
         if (employee.isEmpty()) return null;
+        if (employee.get().getManager() == null) return null;
         if (!isValidDateRange(startDate, endDate)) return null;
 
         TimeOffRequest timeOffRequest = new TimeOffRequest(employee.get(), startDate, endDate);
@@ -62,8 +68,12 @@ public class TimeOffRequestService {
     }
 
     public TimeOffRequest update(long requestID, LocalDateTime startDate, LocalDateTime endDate) {
+        Optional<Employee> employee = employeeRepository.findByUsername(getCurrentUsername());
+        if (employee.isEmpty()) return null;
+
         TimeOffRequest timeOffRequest = timeOffRequestRepository.findById(requestID).orElse(null);
         if (timeOffRequest == null) return null;
+        if (!employee.get().equals(timeOffRequest.getEmployee())) return null;
         if (!isValidDateRange(startDate, endDate)) return null;
         if (timeOffRequest.getRequestStatus() != RequestStatus.PENDING) return null;
 
@@ -74,23 +84,29 @@ public class TimeOffRequestService {
         return save(timeOffRequest);
     }
 
-    public TimeOffRequest judgeRequest(long managerID, long requestID, boolean accept) {
-        Optional<Employee> manager = employeeRepository.findById(managerID);
+    @PreAuthorize("hasAnyAuthority('admin')")
+    public TimeOffRequest judgeRequest(long requestID, boolean accept) {
+        Optional<Employee> manager = employeeRepository.findByUsername(getCurrentUsername());
         if (manager.isEmpty()) return null;
 
         TimeOffRequest timeOffRequest = timeOffRequestRepository.findById(requestID).orElse(null);
         if (timeOffRequest == null) return null;
 
-        timeOffRequest.judge(manager.get(), accept ? RequestStatus.ACCEPTED : RequestStatus.REJECTED);
+        if (!manager.get().equals(timeOffRequest.getManager())) return null;
+
+        timeOffRequest.judge(accept ? RequestStatus.ACCEPTED : RequestStatus.REJECTED);
 
         return timeOffRequestRepository.save(timeOffRequest);
     }
 
     public boolean deleteRequest(long requestID) {
+        Optional<Employee> employee = employeeRepository.findByUsername(getCurrentUsername());
+        if (employee.isEmpty()) return false;
+
         Optional<TimeOffRequest> timeOffRequest = timeOffRequestRepository.findById(requestID);
         if (timeOffRequest.isPresent()) {
-            if (timeOffRequest.get().getRequestStatus() != RequestStatus.PENDING)
-                return false;
+            if (!employee.get().equals(timeOffRequest.get().getEmployee())) return false;
+            if (timeOffRequest.get().getRequestStatus() != RequestStatus.PENDING) return false;
         }
 
         timeOffRequestRepository.deleteById(requestID);
